requirements.txt
discord.py
google-generativeai
playwright
supabase
python-dotenv
flask
requests
pillow
